+++ src/Telegram.Bots/Requests/Admins/UnbanChatMember.cs
@@ -7,13 +7,13 @@ namespace Telegram.Bots.Requests.Admins
   {
     public TChatId ChatId { get; }
 
-    public int UserId { get; }
+    public long UserId { get; }
 
     public bool? OnlyIfBanned { get; init; }
 
     public string Method { get; } = "unbanChatMember";
 
-    protected UnbanChatMember(TChatId chatId, int userId)
+    protected UnbanChatMember(TChatId chatId, long userId)
     {
       ChatId = chatId;
       UserId = userId;
@@ -22,14 +22,14 @@ namespace Telegram.Bots.Requests.Admins
 
   public sealed record UnbanChatMember : UnbanChatMember<long>
   {
-    public UnbanChatMember(long chatId, int userId) : base(chatId, userId) { }
+    public UnbanChatMember(long chatId, long userId) : base(chatId, userId) { }
   }
 
   namespace Usernames
   {
     public sealed record UnbanChatMember : UnbanChatMember<string>
     {
-      public UnbanChatMember(string username, int userId) : base(username, userId) { }
+      public UnbanChatMember(string username, long userId) : base(username, userId) { }
     }
   }
 }
